+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 12.51.0
 
+* `[lock]` New package for working with lock files
 * `[fsutil]` Better errors messages from `ValidatePerms`
 * `[pid]` Code refactoring